Changed random numbers type to Double

+++ R2S1/Java i technologie Web (L)/Lab07/src/lab07_p/Zad01.java
@@ -15,7 +15,7 @@ public class Zad01 {
         Random random = new Random(2137);
         double[] liczby = new double[7];
         for (int i = 0; i < liczby.length; i++) {
-            liczby[i] = random.nextInt(10) + 1;
+            liczby[i] = random.nextDouble(10) + 1;
         }
 
         // Przykład użycia klasy Kalkulator
